requirements.txt
# Core SDR dependencies
numpy>=1.21.0
scipy>=1.7.0

# GUI framework
PyQt5>=5.15.0

# Data handling and storage
h5py>=3.1.0
pandas>=1.3.0

# Device support
pyrtlsdr>=0.2.9

# Configuration and database
pyyaml>=5.4.0
